rename file with test models
